Introduce data "release"
This copies data from a submission into "release" areas, as defined by a
user provided transfer CSV file. Destination keys must being with the
prefix "working", but are otherwise unrestricted.

A versioned manifest of the release transfer is saved in the submission
source bucket under the key:
`release-transfer-manifests/{submission_id}/transfer.{version}`,
where "version" is a formatted datetime, e.g.  "2020-10-26T192530.137917Z"

When using the CLI, releases occur in a single deployment of the SSDS.
This means that objects are copied to a different location of the same
bucket. The Python API can be used to release objects to separate SSDS
deployments.