Introduce data "release" (#213)
This copies data from a submission into "release" areas, as defined by a
user provided transfer CSV file. Destination keys must being with the
prefix "working", but are otherwise unrestricted.

A versioned manifest of the release transfer is saved in the submission
directory under `{submission_pfx}/release-transfer-manifests/{manifest_version}`.
The manifest version is datetime formatted, e.g. "2020-10-26T192530.137917Z".

When using the CLI, releases occur in a single deployment of the SSDS.
This means that objects are copied to a different location of the same
bucket. The Python API can be used to release objects to separate SSDS
deployments.